+++ Src/Private/Get-AbrVbrBackupProxy.ps1
@@ -6,7 +6,7 @@ function Get-AbrVbrBackupProxy {
     .DESCRIPTION
         Documents the configuration of Veeam VBR in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.8.12
+        Version:        0.8.14
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -159,7 +159,7 @@ function Get-AbrVbrBackupProxy {
                                                                 'Processor Model' = $HWCPU[0].Name
                                                                 'Number of CPU Cores' = $HWCPU[0].NumberOfCores
                                                                 'Number of Logical Cores' = $HWCPU[0].NumberOfLogicalProcessors
-                                                                'Physical Memory (GB)' = ConvertTo-FileSizeString -Size $HW.CsTotalPhysicalMemory
+                                                                'Physical Memory (GB)' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $HW.CsTotalPhysicalMemory
                                                             }
                                                             $OutObj += [pscustomobject](ConvertTo-HashToYN $inObj)
 
@@ -193,7 +193,7 @@ function Get-AbrVbrBackupProxy {
                                                                                         'Model' = $Disk.Model
                                                                                         'Serial Number' = $Disk.SerialNumber
                                                                                         'Partition Style' = $Disk.PartitionStyle
-                                                                                        'Disk Size' = ConvertTo-FileSizeString -Size $Disk.Size
+                                                                                        'Disk Size' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $Disk.Size
                                                                                     }
                                                                                     $LocalDiskReport += $TempLocalDiskReport
                                                                                 } catch {
@@ -229,7 +229,7 @@ function Get-AbrVbrBackupProxy {
                                                                                         'Model' = $Disk.Model
                                                                                         'Serial Number' = $Disk.SerialNumber
                                                                                         'Partition Style' = $Disk.PartitionStyle
-                                                                                        'Disk Size' = ConvertTo-FileSizeString -Size $Disk.Size
+                                                                                        'Disk Size' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $Disk.Size
                                                                                     }
                                                                                     $SanDiskReport += $TempSanDiskReport
                                                                                 } catch {
@@ -262,8 +262,8 @@ function Get-AbrVbrBackupProxy {
                                                                                     'Drive Letter' = $HostVolume.DriveLetter
                                                                                     'File System Label' = $HostVolume.FileSystemLabel
                                                                                     'File System' = $HostVolume.FileSystem
-                                                                                    'Size' = ConvertTo-FileSizeString -Size $HostVolume.Size
-                                                                                    'Free Space' = ConvertTo-FileSizeString -Size $HostVolume.SizeRemaining
+                                                                                    'Size' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $HostVolume.Size
+                                                                                    'Free Space' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $HostVolume.SizeRemaining
                                                                                     'Health Status' = $HostVolume.HealthStatus
                                                                                 }
                                                                                 $HostVolumeReport += $TempHostVolumeReport
@@ -611,7 +611,7 @@ function Get-AbrVbrBackupProxy {
                                                                     'Processor Model' = $HWCPU[0].Name
                                                                     'Number of CPU Cores' = $HWCPU[0].NumberOfCores
                                                                     'Number of Logical Cores' = $HWCPU[0].NumberOfLogicalProcessors
-                                                                    'Physical Memory (GB)' = ConvertTo-FileSizeString -Size $HW.CsTotalPhysicalMemory
+                                                                    'Physical Memory (GB)' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $HW.CsTotalPhysicalMemory
                                                                 }
                                                                 $OutObj += [pscustomobject](ConvertTo-HashToYN $inObj)
 
@@ -645,7 +645,7 @@ function Get-AbrVbrBackupProxy {
                                                                                             'Model' = $Disk.Model
                                                                                             'Serial Number' = $Disk.SerialNumber
                                                                                             'Partition Style' = $Disk.PartitionStyle
-                                                                                            'Disk Size' = ConvertTo-FileSizeString -Size $Disk.Size
+                                                                                            'Disk Size' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $Disk.Size
                                                                                         }
                                                                                         $LocalDiskReport += $TempLocalDiskReport
                                                                                     } catch {
@@ -681,7 +681,7 @@ function Get-AbrVbrBackupProxy {
                                                                                             'Model' = $Disk.Model
                                                                                             'Serial Number' = $Disk.SerialNumber
                                                                                             'Partition Style' = $Disk.PartitionStyle
-                                                                                            'Disk Size' = ConvertTo-FileSizeString -Size $Disk.Size
+                                                                                            'Disk Size' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $Disk.Size
                                                                                         }
                                                                                         $SanDiskReport += $TempSanDiskReport
                                                                                     } catch {
@@ -717,8 +717,8 @@ function Get-AbrVbrBackupProxy {
                                                                                         'Drive Letter' = $HostVolume.DriveLetter
                                                                                         'File System Label' = $HostVolume.FileSystemLabel
                                                                                         'File System' = $HostVolume.FileSystem
-                                                                                        'Size' = ConvertTo-FileSizeString -Size $HostVolume.Size
-                                                                                        'Free Space' = ConvertTo-FileSizeString -Size $HostVolume.SizeRemaining
+                                                                                        'Size' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $HostVolume.Size
+                                                                                        'Free Space' = ConvertTo-FileSizeString -RoundUnits $Options.RoundUnits -Size $HostVolume.SizeRemaining
                                                                                         'Health Status' = $HostVolume.HealthStatus
                                                                                     }
                                                                                     $HostVolumeReport += $TempHostVolumeReport