+++ CHANGELOG.md
@@ -7,17 +7,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ##### This project is community maintained and has no sponsorship from Veeam, its employees or any of its affiliates.
 
+## [0.8.14] - 2025-01-30
+
+### Added
+
+- Add support for vMWare Cloud Director backup jobs
+- Add option to set the RoundUnits of the storage usages
+
+### Changed
+
+- Storage units in the report are set to 1 round by default
+- Increase AsBuiltReport.Core to v1.4.2
+
+### Fixed
+
+- Fix [#197](https://github.com/AsBuiltReport/AsBuiltReport.Veeam.VBR/issues/197)
+- Fix [#198](https://github.com/AsBuiltReport/AsBuiltReport.Veeam.VBR/issues/198)
+
 ## [0.8.13] - 2024-12-11
 
 ### Added
 
 - Add EntraID Tenant configuration
   - Add Objects Backup Job information
+  - Add EntraID Tenant information to the Infrastructure diagram
 - Update Malware detection setting
   - Add Signature Detection
 - Update Security & Compliance Best Practice content
 - Add Syslog Event Filter information
 - Add Google Cloud Storage repository information
+- Add VMware Infrastructure information to the Infrastructure diagram
 
 ### Changed